refactor stream handler registry
